issue/#38 move lib/notFound.ts to lib/default.ts
